Moved classes to util package
