+++ PDTTA-core/src/sadl/utils/ProgressBarPrinter.java
@@ -8,7 +8,7 @@
  *
  * You should have received a copy of the GNU General Public License along with SADL.  If not, see <http://www.gnu.org/licenses/>.
  */
-package sadl.modellearner.rtiplus;
+package sadl.utils;
 
 /**
  *
