+++ PDTTA-core/src/sadl/utils/SeqAvgLength.java
@@ -9,7 +9,7 @@
  * You should have received a copy of the GNU General Public License along with SADL.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sadl.modellearner.rtiplus.analysis;
+package sadl.utils;
 
 import java.io.IOException;
 import java.nio.file.Paths;
